requirements.txt
exa-py>=1.1.0
tavily-python>=0.3.0
python-docx>=1.1.2
requests>=2.31.0
tenacity>=8.2.3
